api/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
google-cloud-bigquery
pandas-gbq
db-dtypes
